Change how solr records are harvested
It was using `system_modified_dt`, but it seems active fedora does not
update that field when updating an item's metadata. So use `timestamp`
instead, which is updated by solr everytime the solr record is modified.
This will cause more false positives of items being harvested even
though they did not change, but it will completely remove the false
negatives where a changed object is not harvested.

issue #8